+++ ui-tests/base/pom.xml
@@ -43,5 +43,6 @@
 		<module>com.liferay.ide.ui.theme.deploy.base</module>
 		<module>com.liferay.ide.ui.jsf.deploy.base</module>
 		<module>com.liferay.ide.ui.service.builder.deploy.base</module>
+		<module>com.liferay.ide.ui.hook.deploy.base</module>
 	</modules>
 </project>
