+++ ui-tests/com.liferay.ide.ui.testing-repository/category.xml
@@ -13,4 +13,5 @@
 	<bundle id="com.liferay.ide.ui.theme.deploy.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.jsf.deploy.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.service.builder.deploy.base" version="0.0.0" />
+	<bundle id="com.liferay.ide.ui.hook.deploy.base" version="0.0.0" />
 </site>
