+++ ui-tests/tools/com.liferay.ide.ui.hook/com.liferay.ide.ui.hook.deploy/com.liferay.ide.ui.hook.deploy.tomcat.70/build.properties
