+++ ui-tests/tools/com.liferay.ide.ui.hook/com.liferay.ide.ui.hook.deploy/com.liferay.ide.ui.hook.deploy.tomcat.70/src/com/liferay/ide/ui/hook/tests/DeployHookSdkTests.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.ide.ui.hook.tests;
+
+import com.liferay.ide.ui.hook.deploy.base.DeployHookSdkTomcat7xBase;
+
+import org.junit.Test;
+
+/**
+ * @author Terry Jia
+ * @author Ying Xu
+ */
+public class DeployHookSdkTests extends DeployHookSdkTomcat7xBase {
+
+	@Test
+	public void deployHook() {
+	}
+
+}
