+++ ui-tests/tools/com.liferay.ide.ui.hook/com.liferay.ide.ui.hook.wizard/com.liferay.ide.ui.hook.wizard.create/.gitignore
@@ -1,0 +1,5 @@
+/screenshots
+/.settings
+/.classpath
+/target
+/.project
