+++ ui-tests/tools/com.liferay.ide.ui.hook/com.liferay.ide.ui.hook.wizard/com.liferay.ide.ui.hook.wizard.create/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Liferay IDE UI Hook Wizard Create Tests
+Bundle-SymbolicName: com.liferay.ide.ui.hook.wizard.create.tests
+Bundle-Version: 3.2.0.qualifier
+Bundle-Vendor: Liferay, Inc.
+Require-Bundle: com.liferay.ide.ui.swtbot,
+ com.liferay.ide.ui.liferay,
+ org.eclipse.swtbot.go,
+ org.junit
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
