+++ ui-tests/tools/pom.xml
@@ -48,8 +48,7 @@
 		<module>com.liferay.ide.ui.code.upgrade.tests</module>
 		<module>com.liferay.ide.ui.fragment</module>
 		<module>com.liferay.ide.ui.theme</module>
-		<module>com.liferay.ide.ui.hook.tests</module>
-		<module>com.liferay.ide.ui.hook.deploy.tests</module>
+		<module>com.liferay.ide.ui.hook</module>
 		<module>com.liferay.ide.ui.jsf</module>
 		<module>com.liferay.ide.ui.layouttpl</module>
 		<module>com.liferay.ide.ui.module</module>
